Moved all created charts to src/ivis folder from src/ivis/correlation_charts and added them to ivis.js
