+++ client/src/ivis/CorrelationCharts.scss
@@ -1,4 +1,4 @@
-@import "../../scss/variables";
+@import "../scss/variables";
 
 .headingButtons {
   float: right;
