+++ client/src/ivis/ivis.js
@@ -19,4 +19,8 @@ export * from './Legend';
 export * from './SVG';
 export * from './Records';
 export * from './Selector';
+export * from './ScatterPlot';
+export * from './BubblePlot';
+export * from './HeatmapChart';
+export * from './BarChart';
